EXPERIMENTAL: Graphics: Replace int variables with smaller types
Since most of those are static and many can fit inside a byte or short,
we might as well convert them from int to those types and save a wee
bit of runtime memory while making it clearer the range of values they
are expected to fit into.

+++ src/com/nokia/mid/ui/DirectGraphics.java
@@ -18,24 +18,24 @@ package com.nokia.mid.ui;
 
 public interface DirectGraphics
 {
-	public static final int FLIP_HORIZONTAL = 8192;
-	public static final int FLIP_VERTICAL = 16384;
-	public static final int ROTATE_180 = 180;
-	public static final int ROTATE_270 = 270;
-	public static final int ROTATE_90 = 90;
-	public static final int TYPE_BYTE_1_GRAY = 1;
-	public static final int TYPE_BYTE_1_GRAY_VERTICAL = -1;
-	public static final int TYPE_BYTE_2_GRAY = 2;
-	public static final int TYPE_BYTE_332_RGB = 332;
-	public static final int TYPE_BYTE_4_GRAY = 4;
-	public static final int TYPE_BYTE_8_GRAY = 8;
-	public static final int TYPE_INT_888_RGB = 888;
-	public static final int TYPE_INT_8888_ARGB = 8888;
-	public static final int TYPE_USHORT_1555_ARGB = 1555;
-	public static final int TYPE_USHORT_444_RGB = 444;
-	public static final int TYPE_USHORT_4444_ARGB = 4444;
-	public static final int TYPE_USHORT_555_RGB = 555;
-	public static final int TYPE_USHORT_565_RGB = 565;
+	public static final short FLIP_HORIZONTAL = 8192;
+	public static final short FLIP_VERTICAL = 16384;
+	public static final short ROTATE_180 = 180;
+	public static final short ROTATE_270 = 270;
+	public static final short ROTATE_90 = 90;
+	public static final short TYPE_BYTE_1_GRAY = 1;
+	public static final short TYPE_BYTE_1_GRAY_VERTICAL = -1;
+	public static final short TYPE_BYTE_2_GRAY = 2;
+	public static final short TYPE_BYTE_332_RGB = 332;
+	public static final short TYPE_BYTE_4_GRAY = 4;
+	public static final short TYPE_BYTE_8_GRAY = 8;
+	public static final short TYPE_INT_888_RGB = 888;
+	public static final short TYPE_INT_8888_ARGB = 8888;
+	public static final short TYPE_USHORT_1555_ARGB = 1555;
+	public static final short TYPE_USHORT_444_RGB = 444;
+	public static final short TYPE_USHORT_4444_ARGB = 4444;
+	public static final short TYPE_USHORT_555_RGB = 555;
+	public static final short TYPE_USHORT_565_RGB = 565;
 
 
 	public void drawImage(javax.microedition.lcdui.Image img, int x, int y, int anchor, int manipulation);
